pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "withings-exporter"
version = "0.1.0"
description = "A tool to automatically fetch and export health data from Withings API"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Christoph"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "python-dateutil>=2.8",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
]

[project.scripts]
withings-exporter = "withings_exporter.cli:main"

[tool.setuptools]
packages = ["withings_exporter"]
